LayoutTests/media/modern-media-controls/media-controller/ios/media-controller-ios-layout-traits-modern-compatibility-mode-expected.txt

Testing we use the iOS media controls when in modern compatibility mode on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.shadowRoot(media).querySelector('.media-controls').classList.contains('ios') is true
PASS successfullyParsed is true

TEST COMPLETE
